# swagger_API_DOC.yaml
swagger: "2.0"
info:
  description: "Interview Management API"
  version: "1.0.0"
  title: "Interview Management API"
basePath: "/"
schemes:
- "https"
produces:
- "application/json"
paths:
  /candidate/{candidate_id}:
    get:
      tags:
      - "Get Candidate"
      summary: "Get candidate information"
      description: "Get candidate information from the system and return candidate details"
      operationId: "getCandidate"
      parameters:
      - name: "candidate_id"
        in: "path"
        description: "candidate_id"
        type: "string"
        required: true
      responses:
        200:
          description: "Candidate informations"
          schema:
            $ref: "#/definitions/Candidate"
        default:
          description: "Unexpected Error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "swagger_server.controllers.user_controller"
  /candidate/lists:
    get:
      tags:
      - "Get list of Candidate"
      summary: "Get list of candidate"
      description: "Get list of candidate information from the system and return them"
      operationId: "Candidate_list"
      responses:
        200:
          description: "Candidate informations"
          schema:
            $ref: "#/definitions/Candidate"
        default:
          description: "Unexpected Error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "swagger_server.controllers.user_controller"
  /candidate/create:
    post:
      tags:
      - "Create Candidate"
      summary: "Create Candidate"
      description: "Candidate create endpoint insert and return new candidate information"
      operationId: "candidate_create"
      parameters:
      - name: "candidate"
        in: "body"
        description: "information of candidate."
        required: true
        schema:
          $ref: "#/definitions/Candidate"
      responses:
        201:
          description: "Create new Candidate"
          schema:
            $ref: "#/definitions/Candidate"
        default:
          description: "Unexpected Error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "swagger_server.controllers.user_controller"
  /candidate/update/{candidate_id}:
    put:
      tags:
      - "Update candidate"
      summary: "Update the information of the candidate"
      description: "The Update the information of the candidate endpoint update and return the information of candidate"
      operationId: "candidate_update"
      parameters:
      - name: "candidate_id"
        in: "path"
        description: "candidate_id"
        required: true
        type: "string"
      - name: "candidate"
        in: "body"
        description: "informatin of candidate"
        schema:
          $ref: '#/definitions/Candidate'
      responses:
        200:
          description: "Update the candidate"
          schema:
            $ref: '#/definitions/Candidate'
        default:
          description: "Unexpected Error"
          schema:
            $ref: '#/definitions/Error'
  /candidate/delete/{candidate_id}:
    delete:
      tags:
      - "Remove Candidate"
      summary: "Delete the information of the Candidate"
      description: "The Delete the informatin of the Candidate endpoint Delete and return the id of deleted candidate"
      operationId: "candidate_remove"
      parameters:
      - name: "candidate_id"
        in: "path"
        description: "candidate_id"
        required: true
        type: "string"
      responses:
        404:
          description: "Candidate not found"
          schema:
            $ref: '#/definitions/Error'
  /interviewer/{interviewer_id}:
    get:
      tags:
      - "Get Interviewer"
      summary: "Get interviewer in the system"
      description: "The endpoint will get an interviewer and return interviewer details"
      parameters:
      - name: "interviewer_id"
        in: "path"
        description: "interviewer_id"
        required: true
        type: "string"
      responses:
        200:
          description: "Interviewer details"
        default:
          description: "Unexpected Error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "swagger_server.controllers.user_controller"
  /interviewer/create:
    post:
      tags: 
      - "Create Interviewer"
      summary: "Create Interviewer"
      description: "Interviewer create endpoint insert and return new interviewer information"
      operationId: "interviewer_create"
      parameters:
      - name: "interviewer"
        in: "body"
        description: "information of deliveryman."
        required: true
        schema:
          $ref: "#/definitions/Interviewer"
      responses:
        201:
          description: "Create new Interviewer"
          schema:
            $ref: "#/definitions/Interviewer"
        default:
          description: "Unexpected Error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "swagger_server.controllers.user_controller"
  /interviewer/update/{interviewer_id}:
    put:
      tags:
      - "Update Interviewer"
      summary: "Update the information fo an Interviewer"
      description: "This endpoint will update and interviewer information and return the updated interviewer details"
      operationId: "interviewer_update"
      parameters:
      - name: "interviewer_id"
        in: "path"
        description: "interviewer_id"
        type: "string"
        required: true
      - name: "interviewer"
        in: "body"
        description: "interviewer"
        required: true
        schema:
          $ref: "#/definitions/Interviewer"
      responses:
        200:
          description: "Update the  Interviewer"
          schema: 
            $ref: "#/definitions/Interviewer"
        default:
          description: "Unexpected Error"
          schema: 
            $ref: "#/definitions/Error"
  /interviewer/delete/{interviewer_id}:
    delete:
      tags:
      - "Remove Interviewer"
      summary: "Delete the information of the Interviewer"
      description: "The Delete the informatin of the interviewer endpoint Delete and return the id of deleted interviewer"
      operationId: "interviewer_remove"
      parameters:
      - name: "interviewer_id"
        in: "path"
        description: "interviewer_id"
        required: true
        type: "string"
      responses:
        404:
          description: "Interviewer not found"
          schema:
            $ref: '#/definitions/Error'
  /interviewer/lists:
    get:
      tags:
      - "Get list of interviewers"
      summary: "Get the list of interviewers"
      description: "Get the list of interviewers from the system and return list of interviewers"
      operationId: "interviewer_list"
      responses:
        200:
          description: "interviewers informations"
          schema:
            $ref: "#/definitions/Interviewer"
        default:
          description: "Unexpected Error"
          schema:
            $ref: "#/definitions/Error"
            
  /interview/{interview_id}:
    get:
      tags:
      - "Get interview"
      summary: "Get interview information"
      description: "The Get interview information from the system and return interview details"
      operationId: "interview_get"
      parameters:
      - name: "interview_id"
        in: "path"
        description: "interview_id"
        type: "string"
        required: true
      responses:
        200:
          description: "interview informations"
          schema:
            $ref: "#/definitions/Interview"
        default:
          description: "Unexpected Error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "swagger_server.controllers.user_controller"
  /interview/create:
    post:
      tags:
      - "Create interview"
      summary: "Create interview"
      description: "interview create endpoint insert and return new interview information"
      operationId: "interview_create"
      parameters:
      - name: "interview"
        in: "body"
        description: "information of interview."
        required: true
        schema:
          $ref: "#/definitions/Interview"
      responses:
        201:
          description: "Create new interview"
          schema:
            $ref: "#/definitions/Interview"
        default:
          description: "Unexpected Error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "swagger_server.controllers.user_controller"
  /interview/update/{interview_id}:
    put:
      tags:
      - "Update interview"
      summary: "Update the information of the interview"
      description: "The Update the information of the interview endpoint update and return the information of interview"
      operationId: "interview_update"
      parameters:
      - name: "interview_id"
        in: "path"
        description: "interview_id"
        required: true
        type: "string"
      - name: "interview"
        in: "body"
        description: "informatin of interview"
        schema:
          $ref: '#/definitions/Interview'
      responses:
        200:
          description: "Update the interview"
          schema:
            $ref: '#/definitions/Interview'
        default:
          description: "Unexpected Error"
          schema:
            $ref: '#/definitions/Error'
  /interview/delete/{interview_id}:
    delete:
      tags:
      - "Remove Interview"
      summary: "Delete the information of the Interview"
      description: "The Delete the informatin of the Interview endpoint Delete and return the id of deleted Interview"
      operationId: "interview_remove"
      parameters:
      - name: "interview_id"
        in: "path"
        description: "interview_id"
        required: true
        type: "string"
      responses:
        404:
          description: "interview not found"
          schema:
            $ref: '#/definitions/Error'
  /interview_list:
    get:
      tags:
      - "Get list of interviews"
      summary: "Get the list of interviews"
      description: "Get the list of interviews from the system and return list of interviews"
      operationId: "interview_list"
      responses:
        200:
          description: "interview informations"
          schema:
            $ref: "#/definitions/Interview"
        default:
          description: "Unexpected Error"
          schema:
            $ref: "#/definitions/Error"
  /interview_emptyslot:
    get:
      tags:
      - "Get list of interviews empty slots"
      summary: "Get the liste of interviews empty slots"
      description: "Get the list of interviews empty slot from the system and return list"
      operationId: "interview_slot"
      responses:
        200:
          description: "interview empty slot informations"
          schema:
            $ref: "#/definitions/Interview"
        default:
          description: "Unexpected Error"
          schema:
            $ref: "#/definitions/Error"
  /candidate_assign/{candidate_id}/{interview_id}:
    post:
      tags:
      - " Assign an interview to a candidate"
      summary: "Assign an interview to a candidate"
      description: "this end point permit to assign an avialable intervier slot to a candidate"
      operationId: "candidate_assign"
      parameters:
      - name: "candidate_id"
        in: "path"
        type: "integer"
        description: "ID of the candidate."
        required: true
      - name: "interview_id"
        in: "path"
        description: "ID of the interview to be assign to the candidate."
        type: "integer"
        required: true
      responses:
        200:
          description: "Create new interview"
          schema:
            $ref: "#/definitions/Candidate_assign"
        default:
          description: "Unexpected Error"
          schema:
            $ref: "#/definitions/Error"
  /interviewer_assign/{interviewer_id}/{interview_id}:
    post:
      tags:
      - " Assign an interview to an interviewer"
      summary: "Assign an interview to an interviewer"
      description: "this endpoint permit to assign an avialable intervier slot to an interviewer"
      operationId: "interviwer_assign"
      parameters:
      - name: "interviewer_id"
        in: "path"
        type: "integer"
        description: "ID of the interview to assign to the interviewer."
        required: true
      - name: "interview_id"
        in: "path"
        type: "integer"
        description: "ID of the interview to assign to the interviewer."
        required: true
      responses:
        201:
          description: "Create new interview"
          schema:
            $ref: "#/definitions/Interviewer_assign"
        default:
          description: "Unexpected Error"
          schema:
            $ref: "#/definitions/Error"
  /interview/interviewer/list/{interviewer_id}:
    post:
      tags:
      - " Interviewer interview list"
      summary: "Return the list of  interview of an interviewer"
      description: "this endpoint permit to return list of interview for an interviewer"
      operationId: "interviwer_interviewer_list"
      parameters:
      - name: "interviewer_id"
        in: "path"
        description: "ID of the interviewer to get it list of interview."
        type: "integer"
        required: true
      responses:
        200:
          description: "Create new interview"
          schema:
            $ref: "#/definitions/Interviewer_assign"
        default:
          description: "Unexpected Error"
          schema:
            $ref: "#/definitions/Error"
  /interview/candidate/{candidate_id}:
    post:
      tags:
      - " Candidate interview details"
      summary: "Return the interview  details of a candidate"
      description: "this endpoint permit to return interview details for a candidate "
      operationId: "candidate_interview"
      parameters:
      - name: "candidate_id"
        in: "path"
        description: "ID of the candidate to get the interview details."
        type: "integer"
        required: true
      responses:
        200:
          description: "Create new interview"
          schema:
            $ref: "#/definitions/Interviewer_assign"
        default:
          description: "Unexpected Error"
          schema:
            $ref: "#/definitions/Error"
  
definitions:
  Candidate:
    type: "object"
    properties:
      candidate_id:
        type: "integer"
      firstname: 
        type: "string"
      lastname: 
        type: "string"
      email:
        type: "string"
  Interviewer:
    type: "object"
    properties:
      interviewer_id:
        type: "string"
      firstname: 
        type: "string"
      lastname: 
        type: "string"
      email:
        type: "string"
  Interview:
    type: "object"
    properties:
      interview_id:
        type: "string"
      title: 
        type: "string"
      start_time: 
        type: "string"
        format: date-time
      end_time:
        type: "string"
        format: date-time
  Candidate_assign:
    type: "object"
    properties:
      interview_id:
        type: "string"
      candidate_id: 
        type: "string"
  Interviewer_assign:
    type: "object"
    properties:
      interview_id:
        type: "string"
      candidate_id: 
        type: "string"
  Error:
    type: "object"
    properties:
      code:
        type: "integer"
        format: "int32"
      message:
        type: "string"
      fields:
        type: "string"